+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and run `pre-commit run clk` to replace:
 
 - ✨ Added `yuio.parse.Literal` and support for `typing.Literal`.
 
+- ✨ `yuio.parse.Union` can use choice widget if all of its nested parsers can
+  return options.
+
 ## [2.0.1] - 2026-01-31
 
 - 🐛 Fixed rendering of Json Schema for dicts with string enum keys.